requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Pillow==10.0.1
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-filter==23.3
django-ninja>=0.22.2
pytest
pytest-django
